Seed foods db

+++ db/migrate/20221029081009_add_profile_columns_to_user.rb
@@ -5,11 +5,11 @@ class AddProfileColumnsToUser < ActiveRecord::Migration[7.0]
 
     add_column :users, :gender, :string, default: 'M'
     add_column :users, :activity_type, :string, default: 'active'
-    add_column :users, :height, :numeric
-    add_column :users, :weight, :numeric
+    add_column :users, :height, :float
+    add_column :users, :weight, :float
     add_column :users, :date_of_birth, :date
 
-    add_column :users, :active_metabolic_rate, :numeric
+    add_column :users, :active_metabolic_rate, :float
 
     add_index :users, [:first_name, :last_name], unique: true
   end
