+++ db/migrate/20221029093957_create_foods.rb
@@ -2,10 +2,10 @@ class CreateFoods < ActiveRecord::Migration[7.0]
   def change
     create_table :foods do |t|
       t.string :name
-      t.numeric :calories
-      t.numeric :carbs
-      t.numeric :fat
-      t.numeric :protein
+      t.float :calories
+      t.float :carbs
+      t.float :fat
+      t.float :protein
 
       t.timestamps
     end
