+++ db/migrate/20221029100024_create_food_to_meals.rb
@@ -3,7 +3,7 @@ class CreateFoodToMeals < ActiveRecord::Migration[7.0]
     create_table :food_to_meals do |t|
       t.references :meal, null: false, foreign_key: true
       t.references :food, null: false, foreign_key: true
-      t.numeric :serving_size, null: false
+      t.float :serving_size, null: false
 
       t.timestamps
     end
